Refactor AuthProvider, BackendProvider and SessionProvider.
