+++ app/gui/src/dashboard/App.tsx
@@ -45,14 +45,14 @@ import * as detect from 'enso-common/src/detect'
 
 import * as appUtils from '#/appUtils'
 
-import AuthProvider from '#/providers/AuthProvider'
-import BackendProvider, { useLocalBackend } from '#/providers/BackendProvider'
+import * as authProvider from '#/providers/AuthProvider'
+import { BackendProvider, useLocalBackend } from '#/providers/BackendProvider'
 import InputBindingsProvider from '#/providers/InputBindingsProvider'
 import LocalStorageProvider, * as localStorageProvider from '#/providers/LocalStorageProvider'
 import { useLogger } from '#/providers/LoggerProvider'
 import ModalProvider, * as modalProvider from '#/providers/ModalProvider'
 import * as navigator2DProvider from '#/providers/Navigator2DProvider'
-import SessionProvider from '#/providers/SessionProvider'
+import * as sessionProvider from '#/providers/SessionProvider'
 import * as textProvider from '#/providers/TextProvider'
 
 import VersionChecker from '#/layouts/VersionChecker'
@@ -306,7 +306,7 @@ function AppRouter(props: React.PropsWithChildren<AppRouterProps>) {
 
   return (
     <RouterProvider navigate={navigate}>
-      <SessionProvider
+      <sessionProvider.SessionProvider
         onLogout={() => {
           localStorage.clearUserSpecificEntries()
         }}
@@ -315,15 +315,15 @@ function AppRouter(props: React.PropsWithChildren<AppRouterProps>) {
         registerAuthEventListener={registerAuthEventListener}
       >
         <BackendProvider remoteBackend={remoteBackend} localBackend={localBackend}>
-          <AuthProvider onAuthenticated={onAuthenticated}>
+          <authProvider.AuthProvider onAuthenticated={onAuthenticated}>
             <InputBindingsProvider>
               <LocalBackendPathSynchronizer />
               <VersionChecker />
               {children}
             </InputBindingsProvider>
-          </AuthProvider>
+          </authProvider.AuthProvider>
         </BackendProvider>
-      </SessionProvider>
+      </sessionProvider.SessionProvider>
     </RouterProvider>
   )
 }
