+++ app/gui/src/dashboard/providers/AuthProvider/AuthProvider.tsx
@@ -7,8 +7,9 @@
  */
 import * as React from 'react'
 
-import * as sentry from '@sentry/vue'
+import * as sentry from '@sentry/react'
 import * as reactQuery from '@tanstack/react-query'
+import * as router from 'react-router-dom'
 import * as toast from 'react-toastify'
 import invariant from 'tiny-invariant'
 
@@ -33,93 +34,21 @@ import { EnsoDevtools } from '#/components/Devtools'
 import Page from '#/components/Page'
 import { Result } from '#/components/Result'
 import { useTimeoutCallback } from '#/hooks/timeoutHooks'
+import {
+  featureFlagsForInternalTesting,
+  useFeatureFlag,
+  useSetFeatureFlags,
+} from '#/providers/FeatureFlagsProvider'
 import { isOrganizationId } from '#/services/RemoteBackend'
 import { download } from '#/utilities/download'
 import { getDownloadUrl } from '#/utilities/github'
 import { useMutationCallback } from '#/utilities/tanstackQuery'
 import { unsafeWriteValue } from '#/utilities/write'
-import { useRouterInReact } from '$/providers/react'
 import { Suspense } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
-import {
-  featureFlagsForInternalTesting,
-  useFeatureFlag,
-  useSetFeatureFlags,
-} from './FeatureFlagsProvider'
-
-/** Possible types of {@link BaseUserSession}. */
-// eslint-disable-next-line react-refresh/only-export-components
-export enum UserSessionType {
-  offline = 'offline',
-  partial = 'partial',
-  full = 'full',
-}
-
-/** Properties common to all {@link UserSession}s. */
-interface BaseUserSession extends cognitoModule.UserSession {
-  /** A discriminator for TypeScript to be able to disambiguate between `UserSession` variants. */
-  readonly type: UserSessionType
-}
-
-/**
- * Object containing the currently signed-in user's session data, if the user has not yet set their
- * username.
- *
- * If a user has not yet set their username, they do not yet have an organization associated with
- * their account. Otherwise, this type is identical to the `Session` type. This type should ONLY be
- * used by the `SetUsername` component.
- */
-export interface PartialUserSession extends BaseUserSession {
-  readonly type: UserSessionType.partial
-}
-
-/** Object containing the currently signed-in user's session data. */
-export interface FullUserSession extends BaseUserSession {
-  /** User's organization information. */
-  readonly type: UserSessionType.full
-  readonly user: backendModule.User
-}
-
-/**
- * A user session for a user that may be either fully registered,
- * or in the process of registering.
- */
-export type UserSession = FullUserSession | PartialUserSession
-
-/**
- * Interface returned by the `useAuth` hook.
- *
- * Contains the currently authenticated user's session data, as well as methods for signing in,
- * signing out, etc. All interactions with the authentication API should be done through this
- * interface.
- *
- * See `Cognito` for details on each of the authentication functions.
- */
-interface AuthContextType {
-  readonly authQueryKey: reactQuery.QueryKey
-  readonly setUsername: (username: string) => Promise<boolean>
-  /** @deprecated Never use this function. Prefer particular functions like `setUsername` or `deleteUser`. */
-  readonly setUser: (user: Partial<backendModule.User>) => void
-  readonly deleteUser: () => Promise<boolean>
-  readonly restoreUser: () => Promise<boolean>
-  readonly refetchSession: (
-    options?: reactQuery.RefetchOptions,
-  ) => Promise<reactQuery.QueryObserverResult<UserSession | null>>
-  /**
-   * Session containing the currently authenticated user's authentication information.
-   *
-   * If the user has not signed in, the session will be `null`.
-   */
-  readonly session: UserSession | null
-  /** Return `true` if the user is marked for deletion. */
-  readonly isUserMarkedForDeletion: () => boolean
-  /** Return `true` if the user is deleted completely. */
-  readonly isUserDeleted: () => boolean
-  /** Return `true` if the user is soft deleted. */
-  readonly isUserSoftDeleted: () => boolean
-}
-
-const AuthContext = React.createContext<AuthContextType | null>(null)
+import { AuthContext, useAuth } from './hooks'
+import type { AuthContextType } from './types'
+import { UserSessionType, type FullUserSession, type PartialUserSession } from './types'
 
 /** Query to fetch the user's session data from the backend. */
 function createUsersMeQuery(
@@ -150,7 +79,7 @@ export interface AuthProviderProps {
 }
 
 /** A React provider for the Cognito API. */
-export default function AuthProvider(props: AuthProviderProps) {
+export function AuthProvider(props: AuthProviderProps) {
   const { onAuthenticated, children } = props
 
   const remoteBackend = backendProvider.useRemoteBackend()
@@ -345,49 +274,28 @@ export default function AuthProvider(props: AuthProviderProps) {
 }
 
 /**
- * A React hook that provides access to the authentication context.
- *
- * Only the hook is exported, and not the context, because we only want to use the hook directly and
- * never the context component.
- * @throws {Error} when used outside a {@link AuthProvider}.
- */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useAuth() {
-  const context = React.useContext(AuthContext)
-
-  invariant(context != null, '`useAuth` must be used within an `<AuthProvider />`.')
-
-  return context
-}
-
-/**
  * A React Router layout route containing routes only accessible by users that are logged in.
  */
-export function AnyLoggedInUserLayout({ children }: React.PropsWithChildren) {
+export function AnyLoggedInUserLayout() {
   const { session } = useAuth()
-  const { router } = useRouterInReact()
 
   if (session == null) {
-    void router.push(appUtils.LOGIN_PATH)
-    return
+    return <router.Navigate to={appUtils.LOGIN_PATH} />
   }
 
-  return <>{children}</>
+  return <router.Outlet context={session} />
 }
 
 /** A React Router layout route containing routes only accessible by users that are logged in. */
-export function ProtectedLayout({ children }: React.PropsWithChildren<object>) {
+export function ProtectedLayout() {
   const { session } = useAuth()
-  const { router } = useRouterInReact()
 
   if (session == null) {
-    void router.push(appUtils.LOGIN_PATH)
-    return
+    return <router.Navigate to={appUtils.LOGIN_PATH} />
   }
 
   if (session.type === UserSessionType.partial) {
-    void router.push(appUtils.SETUP_PATH)
-    return
+    return <router.Navigate to={appUtils.SETUP_PATH} />
   }
 
   return (
@@ -396,7 +304,7 @@ export function ProtectedLayout({ children }: React.PropsWithChildren<object>) {
       {/* also it guarantees that the top-level suspense boundary is already resolved */}
       <div data-testid="after-auth-layout" aria-hidden />
 
-      {children}
+      <router.Outlet context={session} />
 
       <Suspense fallback={null}>
         <ErrorBoundary fallbackRender={() => null}>
@@ -411,48 +319,47 @@ export function ProtectedLayout({ children }: React.PropsWithChildren<object>) {
  * A React Router layout route containing routes only accessible by users that are
  * in the process of registering.
  */
-export function SemiProtectedLayout({ children }: React.PropsWithChildren) {
+export function SemiProtectedLayout() {
   const { session } = useAuth()
   const { localStorage } = localStorageProvider.useLocalStorage()
-  const { router } = useRouterInReact()
 
   // The user is not logged in - redirect to the login page.
   if (session == null) {
-    void router.replace(appUtils.LOGIN_PATH)
-    return
+    return <router.Navigate to={appUtils.LOGIN_PATH} replace />
   }
 
   // User is registered, redirect to dashboard or to the redirect path specified during the registration / login.
   if (session.type === UserSessionType.full) {
-    void router.replace(localStorage.consume('loginRedirect') ?? appUtils.DASHBOARD_PATH)
-    return
+    return (
+      <router.Navigate
+        to={localStorage.consume('loginRedirect') ?? appUtils.DASHBOARD_PATH}
+        replace
+      />
+    )
   }
 
   // User is in the process of registration, allow them to complete the registration.
-  return <>{children}</>
+  return <router.Outlet context={session} />
 }
 
 /**
  * A React Router layout route containing routes only accessible by users that are
  * not logged in.
  */
-export function GuestLayout({ children }: React.PropsWithChildren) {
+export function GuestLayout() {
   const { session } = useAuth()
   const { localStorage } = localStorageProvider.useLocalStorage()
-  const { router } = useRouterInReact()
 
   if (session?.type === UserSessionType.partial) {
-    void router.push(appUtils.SETUP_PATH)
-    return
+    return <router.Navigate to={appUtils.SETUP_PATH} />
   } else if (session?.type === UserSessionType.full) {
     const redirectTo = localStorage.get('loginRedirect')
     if (redirectTo != null) {
       localStorage.delete('loginRedirect')
-      void router.push(redirectTo)
+      location.href = redirectTo
       return
     } else {
-      void router.push(appUtils.DASHBOARD_PATH)
-      return
+      return <router.Navigate to={appUtils.DASHBOARD_PATH} />
     }
   } else {
     return (
@@ -460,40 +367,36 @@ export function GuestLayout({ children }: React.PropsWithChildren) {
         {/* This div is used as a flag to indicate that the user is not logged in. */}
         {/* also it guarantees that the top-level suspense boundary is already resolved */}
         <div data-testid="before-auth-layout" aria-hidden />
-        {children}
+        <router.Outlet />
       </>
     )
   }
 }
 
 /** A React Router layout route containing routes only accessible by users that are not deleted. */
-export function NotDeletedUserLayout({ children }: React.PropsWithChildren) {
-  const { isUserMarkedForDeletion } = useAuth()
-  const { router } = useRouterInReact()
+export function NotDeletedUserLayout() {
+  const { session, isUserMarkedForDeletion } = useAuth()
 
   if (isUserMarkedForDeletion()) {
-    void router.push(appUtils.RESTORE_USER_PATH)
+    return <router.Navigate to={appUtils.RESTORE_USER_PATH} />
   } else {
-    return <>{children}</>
+    return <router.Outlet context={session} />
   }
 }
 
 /** A React Router layout route containing routes only accessible by users that are deleted softly. */
-export function SoftDeletedUserLayout({ children }: React.PropsWithChildren) {
-  const { isUserMarkedForDeletion, isUserDeleted, isUserSoftDeleted } = useAuth()
-  const { router } = useRouterInReact()
+export function SoftDeletedUserLayout() {
+  const { session, isUserMarkedForDeletion, isUserDeleted, isUserSoftDeleted } = useAuth()
 
   if (isUserMarkedForDeletion()) {
     const isSoftDeleted = isUserSoftDeleted()
     const isDeleted = isUserDeleted()
     if (isSoftDeleted) {
-      return <>{children}</>
+      return <router.Outlet context={session} />
     } else if (isDeleted) {
-      void router.push(appUtils.LOGIN_PATH)
-      return
+      return <router.Navigate to={appUtils.LOGIN_PATH} />
     } else {
-      void router.push(appUtils.DASHBOARD_PATH)
-      return
+      return <router.Navigate to={appUtils.DASHBOARD_PATH} />
     }
   }
 }
@@ -511,10 +414,9 @@ export interface CloudBrowserDisabledLayoutProps {
 /**
  * Layout that disables the dashboard if the cloud is disabled.
  */
-export function CloudBrowserDisabledLayout(
-  props: React.PropsWithChildren<CloudBrowserDisabledLayoutProps>,
-) {
-  const { children, redirectDelayMs = DEFAULT_REDIRECT_DELAY_MS, redirectPath = '' } = props
+export function CloudBrowserDisabledLayout(props: CloudBrowserDisabledLayoutProps) {
+  const { redirectDelayMs = DEFAULT_REDIRECT_DELAY_MS, redirectPath = '' } = props
+  const { session } = useAuth()
   const { getText } = textProvider.useText()
   const isCloudExecutionEnabled = useFeatureFlag('enableCloudExecution')
   const [isRedirecting, setIsRedirecting] = React.useState(true)
@@ -533,7 +435,7 @@ export function CloudBrowserDisabledLayout(
   })
 
   if (isCloudExecutionEnabled) {
-    return <>{children}</>
+    return <router.Outlet context={session} />
   }
 
   return (
@@ -566,41 +468,4 @@ export function CloudBrowserDisabledLayout(
       </Result>
     </Page>
   )
-}
-
-/**
- * A React context hook returning the user session
- * for a user that has not yet completed registration.
- */
-// eslint-disable-next-line react-refresh/only-export-components
-export function usePartialUserSession() {
-  const { session } = useAuth()
-
-  invariant(session?.type === UserSessionType.partial, 'Expected a partial user session.')
-
-  return session
-}
-
-/** A React context hook returning the user session for a user that may or may not be logged in. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useUserSession() {
-  return useAuth().session
-}
-
-/** A React context hook returning the user session for a user that is fully logged in. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useFullUserSession(): FullUserSession {
-  const { session } = useAuth()
-
-  invariant(session?.type === UserSessionType.full, 'Expected a full user session.')
-
-  return session
-}
-
-/** A React context hook returning the user session for a user that is fully logged in. */
-// eslint-disable-next-line react-refresh/only-export-components
-export function useUser() {
-  const { user } = useFullUserSession()
-
-  return user
 }
