+++ app/gui/src/dashboard/providers/AuthProvider/index.ts
@@ -1,0 +1,3 @@
+export * from './AuthProvider'
+export * from './hooks'
+export * from './types'
