+++ app/gui/src/dashboard/providers/BackendProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './BackendProvider'
+export * from './hooks'
